requirements.txt
streamlit>=1.43.2
websocket-client>=1.8.0
boto3>=1.37.11
python-dotenv>=1.0.0
pydantic>=2.10.6
typing-extensions>=4.8.0
pytest>=7.4.3
mypy>=1.6.0
pytest-cov>=4.1.0 
openai==1.65.4